fixed gdumb

+++ memory.py
@@ -9,7 +9,7 @@ class MemoryContainer(Dataset):
         self.labels = labels
 
     def __getitem__(self, index):
-        return self.data[index], self.labels[index]
+        return self.data[index], self.labels[index].long()
 
     def __len__(self):
         return len(self.data)
